+++ lib/ecto/associations/assoc.ex
@@ -3,9 +3,6 @@ defmodule Ecto.Associations.Assoc do
   This module provides the assoc selector merger and utilities around it.
   """
 
-  alias Ecto.Query.QueryExpr
-  alias Ecto.Query.Util
-
   @doc """
   Transforms a result set based on the assoc selector, loading the associations
   onto their parent model. See `Ecto.Query.select/3`.
@@ -14,25 +11,18 @@ defmodule Ecto.Associations.Assoc do
   def run([], _query), do: []
 
   def run(results, query) do
-    case query.select do
-      %QueryExpr{expr: {:assoc, _, [parent, fields]}} ->
-        merge(results, parent, fields, query)
-      _ ->
+    case query.assocs do
+      [] ->
         results
+      assocs ->
+        merge(results, 0, assocs, query)
     end
   end
 
-  @doc """
-  Decomposes an `assoc(var, fields)` or `var` into `{var, fields}`.
-  """
-  @spec decompose_assoc(Macro.t) :: {Macro.t, [Macro.t]}
-  def decompose_assoc({:&, _, [_]} = var), do: {var, []}
-  def decompose_assoc({:assoc, _, [var, fields]}), do: {var, fields}
-
-  defp merge(rows, var, fields, query) do
+  defp merge(rows, idx, fields, query) do
     # Pre-create rose tree of reflections and accumulator
     # dicts in the same structure as the fields tree
-    refls = create_refls(var, fields, query)
+    refls = create_refls(idx, fields, query.sources)
     accs  = create_accs(fields)
 
     # Populate tree of dicts of associated entities from the result set
@@ -82,21 +72,17 @@ defmodule Ecto.Associations.Assoc do
     {keys, dict, sub_dicts}
   end
 
-  defp create_refls(var, fields, query) do
-    Enum.map(fields, fn {field, nested} ->
-      {child_var, child_fields} = decompose_assoc(nested)
-
-      {_source, model} = Util.find_source(query.sources, var)
-      refl = model.__schema__(:association, field)
-
-      {refl, create_refls(child_var, child_fields, query)}
+  defp create_refls(idx, fields, sources) do
+    Enum.map(fields, fn {field, {child_idx, child_fields}} ->
+      {_source, model} = elem(sources, idx)
+      {model.__schema__(:association, field),
+       create_refls(child_idx, child_fields, sources)}
     end)
   end
 
   defp create_accs(fields) do
-    acc = Enum.map(fields, fn {_field, nested} ->
-      {_, fields} = decompose_assoc(nested)
-      create_accs(fields)
+    acc = Enum.map(fields, fn {_field, {_child_idx, child_fields}} ->
+      create_accs(child_fields)
     end)
 
     {HashSet.new, HashDict.new, acc}