+++ lib/ecto/associations/preloader.ex
@@ -32,7 +32,7 @@ defmodule Ecto.Associations.Preloader do
   end
 
   def run(original, repo, fields, pos) do
-    fields  = normalize(fields)
+    fields  = normalize(fields, fields)
     structs = extract(original, pos)
     structs = Enum.reduce(fields, structs, &do_run(&2, repo, &1))
     unextract(structs, original, pos)
@@ -46,6 +46,7 @@ defmodule Ecto.Associations.Preloader do
     # TODO: Make this use the new Ecto.Model.assoc/2.
     # We just need to prune nils before-hand.
     module = hd(structs).__struct__
+    # TODO: What if reflections is nil?!
     refl   = module.__schema__(:association, field)
     should_sort? = should_sort?(structs, refl)
 
@@ -174,21 +175,21 @@ defmodule Ecto.Associations.Preloader do
 
   ## NORMALIZER ##
 
-  defp normalize(preload) do
-    Enum.map(List.wrap(preload), &normalize_each/1)
+  defp normalize(preload, original) do
+    Enum.map(List.wrap(preload), &normalize_each(&1, original))
   end
 
-  defp normalize_each({atom, list}) when is_atom(atom) do
-    {atom, normalize(list)}
+  defp normalize_each({atom, list}, original) when is_atom(atom) do
+    {atom, normalize(list, original)}
   end
 
-  defp normalize_each(atom) when is_atom(atom) do
+  defp normalize_each(atom, _original) when is_atom(atom) do
     {atom, []}
   end
 
-  defp normalize_each(other) do
-    raise ArgumentError, "invalid preload `#{inspect other}`. preload expects an atom " <>
-                         "a (nested) keyword or a (nested) list of atoms"
+  defp normalize_each(other, original) do
+    raise ArgumentError, "invalid preload `#{inspect other}` in `#{inspect original}`. " <>
+                         "preload expects an atom, a (nested) keyword or a (nested) list of atoms"
   end
 
   ## SORTING ##