+++ test/ecto/query/builder/select_test.exs
@@ -25,19 +25,6 @@ defmodule Ecto.Query.Builder.SelectTest do
             escape(quote do ^x.y end, [])
   end
 
-  test "escape assoc" do
-    assert {Macro.escape(quote do assoc(&0, comments: &1) end), %{}} ==
-            escape(quote do assoc(p, comments: c) end, [p: 0, c: 1])
-
-    assert {Macro.escape(quote do assoc(&0, comments: assoc(&1, author: &2), author: &3) end), %{}} ==
-            escape(quote do assoc(p, comments: assoc(c, author: cu), author: pu) end, [p: 0, c: 1, cu: 2, pu: 3])
-
-    message = "invalid expression `assoc(var, :hey)` inside `assoc/2` selector"
-    assert_raise Ecto.Query.CompileError, message, fn ->
-      escape(quote do assoc(var, :hey) end, [:var])
-    end
-  end
-
   test "only one select is allowed" do
     message = "only one select expression is allowed in query"
     assert_raise Ecto.Query.CompileError, message, fn ->