+++ test/ecto/query/inspect_test.exs
@@ -33,17 +33,17 @@ defmodule Ecto.Query.InspectTest do
   end
 
   test "join" do
-    assert i(from(x in Post, join: y in assoc(x, :comments))) ==
-           ~s{from p in Inspect.Post, join: c in assoc(p, :comments), on: true}
-
-    assert i(from(x in Post, [join: y in Comment, on: x.id == y.id])) ==
+    assert i(from(x in Post, join: y in Comment, on: x.id == y.id)) ==
            ~s{from p in Inspect.Post, join: c in Inspect.Comment, on: p.id == c.id}
 
+    assert i(from(x in Post, join: y in assoc(x, :comments))) ==
+           ~s{from p in Inspect.Post, join: c in assoc(p, :comments)}
+
     assert i(from(x in Post, join: y in assoc(x, :post), join: z in assoc(y, :post))) ==
-           ~s{from p0 in Inspect.Post, join: p1 in assoc(p0, :post), on: true, join: p2 in assoc(p1, :post), on: true}
+           ~s{from p0 in Inspect.Post, join: p1 in assoc(p0, :post), join: p2 in assoc(p1, :post)}
 
     assert i(from(x in Post, left_join: y in assoc(x, :comments))) ==
-           ~s{from p in Inspect.Post, left_join: c in assoc(p, :comments), on: true}
+           ~s{from p in Inspect.Post, left_join: c in assoc(p, :comments)}
   end
 
   test "where" do
@@ -84,6 +84,17 @@ defmodule Ecto.Query.InspectTest do
            ~s{from p in Inspect.Post, lock: "FOOBAR"}
   end
 
+  test "preload" do
+    assert i(from(x in Post, preload: :comments)) ==
+           ~s"from p in Inspect.Post, preload: [:comments]"
+
+    assert i(from(x in Post, join: y in assoc(x, :comments), preload: [comments: y])) ==
+           ~s"from p in Inspect.Post, join: c in assoc(p, :comments), preload: [comments: c]"
+
+    assert i(from(x in Post, join: y in assoc(x, :comments), preload: [comments: {y, post: x}])) ==
+           ~s"from p in Inspect.Post, join: c in assoc(p, :comments), preload: [comments: {c, [post: p]}]"
+  end
+
   test "fragments" do
     value = "foobar"
     assert i(from(x in Post, where: fragment("downcase(?) == ?", x.id, ^value))) ==
@@ -92,23 +103,22 @@ defmodule Ecto.Query.InspectTest do
 
   test "inspect all" do
     string = """
-    from p in Inspect.Post, join: c in assoc(p, :comments), on: true, where: true,
+    from p in Inspect.Post, join: c in assoc(p, :comments), where: true,
     group_by: [p.id], having: true, order_by: [asc: p.id], limit: 1,
-    offset: 1, lock: true, select: 1, preload: :comments
+    offset: 1, lock: true, select: 1, preload: [:likes], preload: [comments: c]
     """
     |> String.rstrip
     |> String.replace("\n", " ")
 
     assert i(from(x in Post, join: y in assoc(x, :comments), where: true, group_by: x.id,
                              having: true, order_by: x.id, limit: 1, offset: 1,
-                             lock: true, select: 1, preload: :comments)) == string
+                             lock: true, select: 1, preload: [:likes, comments: y])) == string
   end
 
   test "to_string all" do
     string = """
     from p in Inspect.Post,
       join: c in assoc(p, :comments),
-      on: true,
       where: true,
       group_by: [p.id],
       having: true,
@@ -117,14 +127,15 @@ defmodule Ecto.Query.InspectTest do
       offset: 1,
       lock: true,
       select: 1,
-      preload: :comments
+      preload: [:likes],
+      preload: [comments: c]
     """
     |> String.rstrip
 
     assert Inspect.Ecto.Query.to_string(
       from(x in Post, join: y in assoc(x, :comments), where: true, group_by: x.id,
                       having: true, order_by: x.id, limit: 1, offset: 1,
-                      lock: true, select: 1, preload: :comments)
+                      lock: true, select: 1, preload: [:likes, comments: y])
     ) == string
   end